refactor(app): moving Message data class to domain package
