+++ .idea/misc.xml
@@ -17,6 +17,7 @@
         <entry key="../../../../../layout/compose-model-1628713326162.xml" value="0.33" />
         <entry key="../../../../../layout/compose-model-1628713436839.xml" value="0.33" />
         <entry key="../../../../../layout/compose-model-1628713436844.xml" value="0.9097938144329897" />
+        <entry key="../../../../../layout/compose-model-1628718541076.xml" value="0.16666666666666666" />
       </map>
     </option>
   </component>
