+++ app/src/main/java/com/powilliam/android/chatting/ui/composables/MessageCardList.kt
@@ -12,6 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.powilliam.android.chatting.domain.models.Message
 import com.powilliam.android.chatting.ui.ChattingTheme
 
 @OptIn(ExperimentalFoundationApi::class)
